try catch using eval
Created a subroutine named try that executes code passed to it.  To pass code to try by enclosing it in the customary curly braces, I prototype try to accept a reference to an anonymous subroutine (which I create with curly braces). 

To execute the code in the try block then, I just call the anonymous subroutine that was passed to try and report any eval errors, as in this where a division by zero is attempted.