Print Perl CGI ENV variables
How can I print all of the Perl CGI environment variables? I'd like to see what CGI environment variables are available, and what their values are. It can be used to help debug an HTTP server, and other times I look at it as a reminder of how to work with associative arrays (also called hashes in the latest Perl lingo).